requirements.txt
django-ipgeolocation
requests
django-ratelimit
celery
redis

